fix(goenv): remove GO{PATH,ROOT}/bin from PATH
When we start a terminal in somewhere .go-version existed, go env will
setup GOPATH and GOROOT when we run `goenv init` in zshenv.

These two environment variables will then be appended to PATH, which
will not update even after we call goenv shell.

So it seems a good idea to remove them from PATH, as they are not quite
useful as I was excepted.